Make length of training iteration dependent on world dim
Previously, iteration was ended when the snake died or 5
states had been collected. This does not scale well with
increasing / varying sizes of the game world. Try to use
the sum of these dimensions instead (that way, the snake
is guaranteed to be able to cross the entire map and reach
at least one food item per iteration.)