Fix transitions and add final state
Transitions were created across batches so that
the next_state of one batch's last state was actually
the first state of the next batch.

For actions that cause the snake to die, use the immediate
reward only - as there is no future state, there is also no
expected future reward.